[inductor] Generate indirect_indexing checks even if optimized out
Previously `gen_assert_indirect_indexing` was only called on the index
expressions passed to `ops.load` and `ops.store` which means if the
variable is optimized out during lowering, we never generate the
assert. This instead makes `ops.indirect_indexing` eagerly generate
the assert statement, whether or not it will be used.

ghstack-source-id: b0ae213f3a152775ee7f3d212d97338a765d5513
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100895